+++ packages/uploader/src/react/uploader.tsx
@@ -11,7 +11,6 @@ import {
   useEffect,
   useImperativeHandle,
   lazy,
-  useLayoutEffect,
 } from 'react';
 import get from 'lodash/get';
 import { useTheme } from '@mui/material/styles';
@@ -83,7 +82,7 @@ const uploaderDashboardId = 'uploader-dashboard';
 const isDebug = localStorage.getItem('uppy_debug');
 
 const getPluginList = (props: any) => {
-  const { apiPathProps, availablePluginMap = {}, uploadedProps, resourcesProps, theme } = props;
+  const { apiPathProps, availablePluginMap = {}, uploadedProps, resourcesProps, imageEditorProps = {}, theme } = props;
 
   const { companionUrl } = getUploaderEndpoint(apiPathProps);
 
@@ -104,6 +103,8 @@ const getPluginList = (props: any) => {
       plugin: ImageEditor, // use image editor
       options: {
         quality: 1,
+        // docs: https://uppy.io/docs/image-editor/#options
+        ...imageEditorProps,
       },
       alwayUse: true,
     },
@@ -212,6 +213,7 @@ const getPluginList = (props: any) => {
       plugin: PrepareUpload,
       options: {
         companionUrl,
+        cropperOptions: imageEditorProps?.cropperOptions || null,
       },
       alwayUse: true,
     },
@@ -880,7 +882,7 @@ const Uploader = forwardRef((props: UploaderProps & IframeHTMLAttributes<HTMLIFr
                 width: '70%',
               },
             },
-            '& .uppy-Dashboard-browse, & .uppy-DashboardContent-addMore, & .uppy-DashboardContent-back, & .uppy-StatusBar-actionBtn--done':
+            '& .uppy-Dashboard-browse, & .uppy-DashboardContent-addMore, & .uppy-DashboardContent-back, & .uppy-StatusBar-actionBtn--done, & .uppy-DashboardContent-save, & .uppy-StatusBar-actionBtn--upload':
               {
                 color: `${theme?.palette?.primary?.main}`,
                 transition: 'all 0.3s ease-in-out',